Added first spec
