+++ script/test
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+rspec $1
